refactor : 명세서에 맞게 api 수정
- 오늘 만들어진 api 명세서에 따라 일부 리펙토링 했습니다.

related #13
